backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo[srv]==4.6.2
python-dotenv==1.0.1
mercadopago==2.3.0
httpx==0.27.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.5
pydantic[email]==2.6.3
python-multipart==0.0.9
